Modularizamos las distintas rondas que aparecen en la seccion de /tournament-details, mas especificamente en la seccion de brackets donde aparecen las diferentes rondas en las que ocurrieron las partidas, en este caso, ahora hay un archivo para cada uno de ellos
