+++ app/src/components/Componentes Competicion/CuadrosBracket.js
@@ -1,7 +1,12 @@
 import React, { useState, useRef } from "react";
 import MatchModal from "../Componentes Detalles/MatchModal"; // Importa el componente que muestra el modal del partido
 import MatchDetailsPopup from "../Componentes Detalles/MatchDetailsPopup"; // Importa el componente que muestra los detalles emergentes del partido
-
+import Ronda32 from "./Ronda32BracketComponent";
+import Dieciseisavos from "./DieciseisavosComponent";
+import Octavos from "./OctavosComponent";
+import Cuartos from "./CuartosComponent";
+import Semis from "./SemisComponent";
+import Final from "./FinalComponent";
 const CuadrosBracket = ({ positions }) => {
   // Estado para mantener el equipo resaltado
   const [highlightedTeam, setHighlightedTeam] = useState(null);
@@ -27,8 +32,6 @@ const CuadrosBracket = ({ positions }) => {
   // Estado que guarda el partido seleccionado para mostrar sus detalles
   const [selectedMatch, setSelectedMatch] = useState(null);
 
-
-  
 
   
   // Función para manejar el clic sobre un equipo, estableciendo el equipo seleccionado en el modal
@@ -57,60 +60,6 @@ const CuadrosBracket = ({ positions }) => {
   // Función para manejar cuando se suelta el mouse (fin del arrastre)
   const handleMouseUp = () => setIsDragging(false);
 
-  // Función para manejar el clic en un partido, mostrando los detalles del partido
-  const handleMatchClick = (matchNumber, team1, team2) => {
-    setSelectedMatch({
-      matchNumber,
-      team1,
-      team2
-    });
-    setShowMatchDetails(true); // Mostrar los detalles del partido seleccionado
-  };
-
-  // Función para determinar el estilo de la tarjeta del equipo (para mostrar su cuadro)
-  const getCardStyle = (team) => {
-    return `bg-gray-800 ${highlightedTeam === team ? "border-gray-600 shadow-lg" : "border-gray-700"}`; // Si es el equipo resaltado, aplicar borde y sombra especial
-  };
-
-  // Estilos específicos para el ganador (finalista)
-
-
-  // Lista de nombres de equipos
-  const teamNames = [
-    "Dragons", "Warriors", "Titans", "Phoenix", "Sharks", "Wolves", "Eagles", "Lions",
-    "Bears", "Panthers", "Hawks", "Tigers", "Bulls", "Raptors", "Knights", "Spartans",
-    "Vikings", "Samurais", "Ninjas", "Pirates", "Gladiators", "Raiders", "Giants", "Cyclones",
-    "Storm", "Thunder", "Lightning", "Blaze", "Inferno", "Tornadoes", "Avalanche", "Quakes"
-  ];
-
-  // Equipos ganadores de la primera ronda (16avos de final)
-  const dieciseisavosWinners = [
-    "Dragons", "Titans", "Sharks", "Eagles", "Bears", "Hawks", "Bulls", "Knights",
-    "Vikings", "Ninjas", "Gladiators", "Giants", "Storm", "Lightning", "Inferno", "Avalanche"
-  ];
-
-  // Equipos ganadores de los octavos de final
-  const octavosWinners = [
-    "Dragons", "Sharks", "Bears", "Bulls", "Vikings", "Gladiators", "Storm", "Inferno"
-  ];
-
-  // Equipos ganadores de los cuartos de final
-  const cuartosWinners = [
-    "Dragons", "Bears", "Vikings", "Storm"
-  ];
-
-  // Equipos ganadores de las semifinales
-  const semifinalWinners = [
-    "Dragons", "Storm"
-  ];
-
-  // Ganador final, subcampeón y tercer lugar
-  const finalWinner = "Dragons";
-  const secondPlaceWinner = "Storm";
-  const thirdPlaceWinner = "Bears";
-
-  // Descripción de las rondas (partidas)
-  const partidas = ['16 partidas','8 partidas','4 partidas','2 partidas','1 partida'];
 
 
   return (
@@ -125,22 +74,7 @@ const CuadrosBracket = ({ positions }) => {
         }
       `}</style>
 
-      <div className="flex text-xs font-semibold mb-6 relative z-10 justify-center">
-        {['RONDA DE 32', 'OCTAVOS DE FINAL', 'CUARTOS DE FINAL', 'SEMIFINALES', 'FINAL Y 3ER PUESTO'].map((round, i) => (
-          <div key={`round-header-${i}`} className="w-1/5 px-1">
-            <div className="py-2 text-center rounded flex flex-col">
-              <p className="text-lg">
-              {round}
-              </p>
-
-              <p className="text-blue-500 text-sm">
-              {partidas[i]}
-              </p>
-              
-            </div>
-          </div>
-        ))}
-      </div>
+      <Ronda32></Ronda32>
 
       <div
         ref={bracketRef}
@@ -151,293 +85,14 @@ const CuadrosBracket = ({ positions }) => {
         onMouseLeave={handleMouseUp}
       >
         <div className="absolute inset-0 bg-[#1c1c1c] w-full h-full -z-10"></div>
-        <div className="w-1/4 pr-2"> {/* Dieciseisavos */}
-          {Array.from({ length: 16 }).map((_, i) => (
-            <div key={`dieciseisavos-${i}`} className="mb-8 relative ml-8" style={{ top: positions.dieciseisavos[i].top }}>
-              <button
-                className="absolute top-[-1.5rem] left-0 text-sm font-bold text-blue-500"
-              >
-                Partida {i + 1}
-              </button>
-              <div className={`${getCardStyle(teamNames[i * 2])} text-sm rounded overflow-hidden shadow-lg w-[12rem]`}>
-                <div className={`border-b border-gray-700 py-2 px-3 flex justify-between items-center transition-colors duration-150 `}
-                  
-                >
-                  <div className="flex items-center font-bold text-white">
-                    <div className="w-6 h-6 rounded-full bg-gray-600 mr-2 overflow-hidden flex-shrink-0">
-                      <img src={`/logos/${teamNames[i * 2].toLowerCase()}.png`} alt={`${teamNames[i * 2]} logo`} className="w-full h-full object-cover" />
-                    </div>
-                    <p className="text-yellow-500">{teamNames[i * 2]}</p> 
-                  </div>
-                  <span >{i % 2 === 0 ? 2 : 1}</span>
-                </div>
-                <div className={`py-2 px-3 flex justify-between items-center transition-colors duration-150 $`}
-                  
-                >
-                  <div className="flex items-center font-bold text-white">
-                    <div className="w-6 h-6 rounded-full bg-gray-600 mr-2 overflow-hidden flex-shrink-0">
-                      <img src={`/logos/${teamNames[i * 2 + 1].toLowerCase()}.png`} alt={`${teamNames[i * 2 + 1]} logo`} className="w-full h-full object-cover" />
-                    </div>
-                    <p className="text-blue-500">{teamNames[i * 2 + 1]}</p>
-                  </div>
-                  <span>{i % 2 === 0 ? 1 : 2}</span>
-                </div>
-                <div className="bg-[#1c1c1c] px-3 py-2 flex justify-center">
-                  <button 
-                    onClick={() => handleMatchClick(i + 1, teamNames[i * 2], teamNames[i * 2 + 1])}
-                    className="bg-blue-600 hover:bg-blue-700 text-white text-xs px-3 py-1 rounded-full transition-colors duration-200"
-                  >
-                    Ver Detalles
-                  </button>
-                </div> 
-              </div>
-            </div>
-          ))}
-        </div>
         
-        <div className="w-1/5 pr-2"> {/* Octavos de Final */}
-          {Array.from({ length: 8 }).map((_, i) => (
-            <div key={`octavos-${i}`} className="mb-8 relative ml-8" style={{ top: positions.octavos[i].top }}>
-              <button
-                className="absolute top-[-1.5rem] left-0 text-sm font-bold text-blue-500"
-              >
-                Partida {i + 17}
-              </button>
-              <div className={`${getCardStyle(dieciseisavosWinners[i * 2])} text-sm rounded overflow-hidden shadow-lg w-[12rem]`}>
-                <div className={`border-b border-gray-700 py-2 px-3 flex justify-between items-center transition-colors duration-150 `}
-                  
-                >
-                  <div className="flex items-center font-bold text-white">
-                    <div className="w-6 h-6 rounded-full bg-gray-600 mr-2 overflow-hidden flex-shrink-0">
-                      <img src={`/logos/${dieciseisavosWinners[i * 2].toLowerCase()}.png`} alt={`${dieciseisavosWinners[i * 2]} logo`} className="w-full h-full object-cover" />
-                    </div>
-                    <p className="text-yellow-500">{dieciseisavosWinners[i * 2]}</p>
-                    
-                  </div>
-                  <span>{i % 2 === 0 ? 2 : 1}</span>
-                </div>
-                <div className={`py-2 px-3 flex justify-between items-center transition-colors duration-150 `}
-                 
-                >
-                  <div className="flex items-center font-bold text-white">
-                    <div className="w-6 h-6 rounded-full bg-gray-600 mr-2 overflow-hidden flex-shrink-0">
-                      <img src={`/logos/${dieciseisavosWinners[i * 2 + 1].toLowerCase()}.png`} alt={`${dieciseisavosWinners[i * 2 + 1]} logo`} className="w-full h-full object-cover" />
-                    </div>
-                    <p className="text-blue-500">{dieciseisavosWinners[i * 2 + 1]}</p>
-                  </div>
-                  <span>{i % 2 === 0 ? 1 : 2}</span>
-                </div>
-                <div className="bg-[#1c1c1c] px-3 py-2 flex justify-center">
-                  <button 
-                    onClick={() => handleMatchClick(i + 17, dieciseisavosWinners[i * 2], dieciseisavosWinners[i * 2 + 1])}
-                    className="bg-blue-600 hover:bg-blue-700 text-white text-xs px-3 py-1 rounded-full transition-colors duration-200"
-                  >
-                    Ver Detalles
-                  </button>
-                </div>
-              </div>
-              <div className="absolute left-[-1rem] top-[50%] w-[1rem] h-[1px] bg-white"></div>
-              <div className="absolute left-[-1rem] top-[50%] w-[1px] h-[3rem] bg-white"></div>
-              <div className="absolute left-[-1rem] bottom-[50%] w-[1px] h-[8rem] bg-white"></div>
-              <div className="absolute left-[-1rem] top-[calc(50%+3rem)] w-[1px] h-[2rem] bg-white"></div>
-              <div className="absolute left-[-2rem] top-[calc(50%+5rem)] w-[1rem] h-[1px] bg-white"></div>
-              <div className="absolute left-[-2rem] bottom-[calc(50%+8rem)] w-[1rem] h-[1px] bg-white"></div>
-            </div>
-          ))}
-        </div>
+        <Dieciseisavos positions = {positions}></Dieciseisavos>
+        <Octavos positions = {positions}></Octavos>
+        <Cuartos positions = {positions}></Cuartos>
+        <Semis positions = {positions}></Semis>
+        <Final positions = {positions}></Final>
         
-        <div className="w-1/5 pr-2"> {/* Cuartos de Final */}
-          {Array.from({ length: 4 }).map((_, i) => (
-            <div key={`cuartos-${i}`} className="mb-16 relative ml-8" style={{ top: positions.cuartos[i].top }}>
-              <button
-                className="absolute top-[-1.5rem] left-0 text-sm font-bold text-blue-500"
-              >
-                Partida {i + 25}
-              </button>
-              <div className={`${getCardStyle(octavosWinners[i * 2])} text-sm rounded overflow-hidden shadow-lg w-[12rem]`}>
-                <div className={`border-b border-gray-700 py-2 px-3 flex justify-between items-center transition-colors duration-150 `}
-                  
-                >
-                  <div className="flex items-center font-bold text-white">
-                    <div className="w-6 h-6 rounded-full bg-gray-600 mr-2 overflow-hidden flex-shrink-0">
-                      <img src={`/logos/${octavosWinners[i * 2].toLowerCase()}.png`} alt={`${octavosWinners[i * 2]} logo`} className="w-full h-full object-cover" />
-                    </div>
-                    <p className="text-yellow-500">{octavosWinners[i * 2]}</p>
-                   
-                  </div>
-                  <span>{i % 2 === 0 ? 2 : 1}</span>
-                </div>
-                <div className={`py-2 px-3 flex justify-between items-center transition-colors duration-150 $`}
-                  
-                >
-                  <div className="flex items-center font-bold text-white">
-                    <div className="w-6 h-6 rounded-full bg-gray-600 mr-2 overflow-hidden flex-shrink-0">
-                      <img src={`/logos/${octavosWinners[i * 2 + 1].toLowerCase()}.png`} alt={`${octavosWinners[i * 2 + 1]} logo`} className="w-full h-full object-cover" />
-                    </div>
-                    
-                    <p className="text-blue-500">{octavosWinners[i * 2 + 1]}</p>
-                  </div>
-                  <span>{i % 2 === 0 ? 1 : 2}</span>
-                </div>
-                <div className="bg-[#1c1c1c] px-3 py-2 flex justify-center">
-                  <button 
-                    onClick={() => handleMatchClick(i + 25, octavosWinners[i * 2], octavosWinners[i * 2 + 1])}
-                    className="bg-blue-600 hover:bg-blue-700 text-white text-xs px-3 py-1 rounded-full transition-colors duration-200"
-                  >
-                    Ver Detalles
-                  </button>
-                </div>
-              </div>
-              <div className="absolute left-[-1rem] top-[50%] w-[1rem] h-[1px] bg-white"></div>
-              <div className="absolute left-[-1rem] top-[50%] w-[1px] h-[12rem] bg-white"></div>
-              <div className="absolute left-[-1rem] bottom-[50%] w-[1px] h-[16rem] bg-white"></div>
-              <div className="absolute left-[-1rem] top-[calc(50%+5rem)] w-[1px] h-[5rem] bg-white"></div>
-              <div className="absolute left-[-2rem] top-[calc(50%+12rem)] w-[1rem] h-[1px] bg-white"></div>
-              <div className="absolute left-[-2rem] bottom-[calc(50%+16rem)] w-[1rem] h-[1px] bg-white"></div>
-            </div>
-          ))}
-        </div>
         
-        <div className="w-1/4 pr-2"> {/* Semifinales */}
-          {Array.from({ length: 2 }).map((_, i) => (
-            <div key={`semifinales-${i}`} className="mb-16 relative ml-8" style={{ top: positions.semifinales[i].top }}>
-              <button
-                className="absolute top-[-1.5rem] left-0 text-sm font-bold text-blue-500"
-              >
-                Partida {i + 29}
-              </button>
-              <div className={`${getCardStyle(cuartosWinners[i * 2])} text-sm rounded overflow-hidden shadow-lg w-[12rem]`}>
-                <div className={`border-b border-gray-700 py-2 px-3 flex justify-between items-center transition-colors duration-150 `}
-
-                 
-                >
-                  <div className="flex items-center font-bold text-white">
-                    <div className="w-6 h-6 rounded-full bg-gray-600 mr-2 overflow-hidden flex-shrink-0">
-                      <img src={`/logos/${cuartosWinners[i * 2].toLowerCase()}.png`} alt={`${cuartosWinners[i * 2]} logo`} className="w-full h-full object-cover" />
-                    </div>
-                    <p className="text-yellow-500">{cuartosWinners[i * 2]}</p>
-                  </div>
-                  <span>2</span>
-                </div>
-                <div className={`py-2 px-3 flex justify-between items-center transition-colors duration-150 $`}
-                  
-                >
-                  <div className="flex items-center font-bold text-white">
-                    <div className="w-6 h-6 rounded-full bg-gray-600 mr-2 overflow-hidden flex-shrink-0">
-                      <img src={`/logos/${cuartosWinners[i * 2 + 1].toLowerCase()}.png`} alt={`${cuartosWinners[i * 2 + 1]} logo`} className="w-full h-full object-cover" />
-                    </div>
-                    <p className="text-blue-500">{octavosWinners[i * 2 + 1]}</p>
-                  </div>
-                  <span>1</span>
-                </div>
-                <div className="bg-[#1c1c1c] px-3 py-2 flex justify-center">
-                  <button 
-                    onClick={() => handleMatchClick(i + 29, cuartosWinners[i * 2], cuartosWinners[i * 2 + 1])}
-                    className="bg-blue-600 hover:bg-blue-700 text-white text-xs px-3 py-1 rounded-full transition-colors duration-200"
-                  >
-                    Ver Detalles
-                  </button>
-                </div>
-              </div>
-              <div className="absolute left-[-1rem] top-[50%] w-[1rem] h-[1px] bg-white"></div>
-              <div className="absolute left-[-1rem] top-[50%] w-[1px] h-[27rem] bg-white"></div>
-              <div className="absolute left-[-1rem] bottom-[50%] w-[1px] h-[30rem] bg-white"></div>
-              <div className="absolute left-[-1rem] top-[calc(50%+9rem)] w-[1px] h-[9rem] bg-white"></div>
-              <div className="absolute left-[-2rem] top-[calc(50%+27rem)] w-[1rem] h-[1px] bg-white"></div>
-              <div className="absolute left-[-2rem] bottom-[calc(50%+30rem)] w-[1rem] h-[1px] bg-white"></div>
-            </div>
-          ))}
-        </div>
-        <div className="w-1/4 pr-2"> {/* Final */}
-          <div className="mt-20 mb-16 relative ml-8" style={{ top: positions.final.top }}>
-            <button
-              className="absolute top-[-1.5rem] left-0 text-sm font-bold text-blue-500"
-            >
-              Partida 31
-            </button>
-            <div className="bg-gray-800 text-sm rounded overflow-hidden shadow-lg w-[12rem]">
-              <div className={`border-b border-gray-700 py-2 px-3 flex justify-between items-cente bg-opacity-80 transition-colors duration-150  text-gold`}>
-                <div className="flex items-center font-bold text-white">
-                  <div className="w-6 h-6 rounded-full bg-gray-600 mr-2 overflow-hidden flex-shrink-0">
-                    <img src={`/logos/${finalWinner.toLowerCase()}.png`} alt={`${finalWinner} logo`} className="w-full h-full object-cover" />
-                  </div>
-                  <p className="text-yellow-500">{finalWinner}</p>
-                  
-                </div>
-                <span>2</span>
-              </div>
-              <div className={`py-2 px-3 flex justify-between items-center transition-colors duration-150 `}
-                
-              >
-                <div className="flex items-center font-bold text-white">
-                  <div className="w-6 h-6 rounded-full bg-gray-600 mr-2 overflow-hidden flex-shrink-0">
-                    <img src={`/logos/${secondPlaceWinner.toLowerCase()}.png`} alt={`${secondPlaceWinner} logo`} className="w-full h-full object-cover" />
-                  </div>
-                  <p className="text-blue-500">{secondPlaceWinner}</p>
-                  
-                </div>
-                <span >1</span>
-              </div>
-              <div className="bg-[#1c1c1c] px-3 py-2 flex justify-center">
-                <button 
-                  onClick={() => handleMatchClick(31, finalWinner, secondPlaceWinner)}
-                  className="bg-blue-600 hover:bg-blue-700 text-white text-xs px-3 py-1 rounded-full transition-colors duration-200"
-                >
-                  Ver Detalles
-                </button>
-              </div>
-            </div>
-            <div className="absolute left-[-1rem] top-[50%] w-[1rem] h-[1px] bg-white"></div>
-            <div className="absolute left-[-1rem] top-[50%] w-[1px] h-[30rem] bg-white"></div>
-            <div className="absolute left-[-1rem] bottom-[50%] w-[1px] h-[58rem] bg-white"></div>
-            <div className="absolute left-[-1rem] top-[calc(50%+17rem)] w-[1px] h-[40rem] bg-white"></div>
-            <div className="absolute left-[-2rem] top-[calc(50%+57rem)] w-[1rem] h-[1px] bg-white"></div>
-            <div className="absolute left-[-2rem] bottom-[calc(50%+58rem)] w-[1rem] h-[1px] bg-white"></div>
-          </div>
-          <div className="mt-44 mb-16 relative ml-8" style={{ top: positions.tercerPuesto.top }}>
-            <button
-              className="absolute top-[-1.5rem] left-0 text-sm font-bold text-blue-500"
-            >
-              Partida 32
-            </button>
-            <div className="bg-gray-800 text-sm rounded overflow-hidden shadow-lg w-[12rem]">
-              <div className={`border-b border-gray-700 py-2 px-3 flex justify-between items-center transition-colors duration-150 `}
-                
-              >
-                <div className="flex items-center font-bold text-white">
-                  <div className="w-6 h-6 rounded-full bg-gray-600 mr-2 overflow-hidden flex-shrink-0">
-                    <img src={`/logos/${thirdPlaceWinner.toLowerCase()}.png`} alt={`${thirdPlaceWinner} logo`} className="w-full h-full object-cover" />
-                  </div>
-                  <p className="text-yellow-500">{secondPlaceWinner}</p>
-                  
-                </div>
-                <span >2</span>
-              </div>
-              <div className={`py-2 px-3 flex justify-between items-center transition-colors duration-150 `}
-                
-              >
-                <div className="flex items-center font-bold text-white">
-                  <div className="w-6 h-6 rounded-full bg-gray-600 mr-2 overflow-hidden flex-shrink-0">
-                    <img src={`/logos/${cuartosWinners[3].toLowerCase()}.png`} alt={`${cuartosWinners[3]} logo`} className="w-full h-full object-cover" />
-                  </div>
-                  
-                  <p className="text-blue-500">{thirdPlaceWinner}</p>
-                  
-                </div>
-                <span>1</span>
-              </div>
-              <div className="bg-[#1c1c1c] px-3 py-2 flex justify-center">
-                <button 
-                  onClick={() => handleMatchClick(32, thirdPlaceWinner, cuartosWinners[3])}
-                  className="bg-blue-600 hover:bg-blue-700 text-white text-xs px-3 py-1 rounded-full transition-colors duration-200"
-                >
-                  Ver Detalles
-                </button>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
 
       {/* Modal */}
       {modalTeam && (
@@ -476,6 +131,7 @@ const CuadrosBracket = ({ positions }) => {
           onClose={() => setShowMatchDetails(false)}
         />
       )}
+    </div>
     </div>
   );
 };
