+++ app/src/components/Header.js
@@ -61,8 +61,8 @@ const Header = () => {
         </NavLink>
       </nav>
       <div className="flex items-center gap-2 ml-auto">
-        <a href="/register" className="text-white no-underline px-2 py-2 rounded transition-all duration-300 hover:-translate-y-1 hover:shadow-md border-2 border-white bg-transparent">Registrarse</a>
-        <a href="/login" className="text-white no-underline px-2 py-2 rounded transition-all duration-300 hover:-translate-y-1 hover:shadow-md bg-blue-600 ml-4">Iniciar Sesión</a>
+        <a href="/register" className="text-white no-underline px-2 py-2 rounded transition-all duration-300 hover:-translate-y-1 hover:shadow-md bg-gradient-to-r from-[#005f99] to-[#1AA9FF]">Registrarse</a>
+        <a href="/login" className="text-white no-underline px-2 py-2 rounded transition-all duration-300 hover:-translate-y-1 hover:shadow-md bg-gradient-to-r from-[#005f99] to-[#1AA9FF] ml-4">Iniciar Sesión</a>
       </div>
     </header>
   );
