+++ app/src/components/componentesInicio/FuncionaInicio.js
@@ -1,0 +1,59 @@
+import React from "react";
+import Competiciones from "@imgs/Header_Competiciones.jpg";
+import Contenido from "@imgs/Headers_Contenido.jpg";
+import Misiones from "@imgs/Header_Misiones.jpg";
+const FuncionaInicio = () => {
+    return (
+        <div className="text-white py-[8vh] px-6 sm:px-12 bg-[#1AA9FF]"> {/* Reduced top padding */}
+          <div>
+            <h1 className="text-4xl sm:text-5xl lg:text-6xl font-bold text-center mb-16 sm:mb-20 tracking-wide">Descubre lo que Oxyclub tiene para ti</h1>
+            <div className="grid place-items-center gap-[6rem] grid-cols-1 md:grid-cols-3 w-full">
+              {[
+                {
+                  icon: "fas fa-trophy",
+                  title: "Competiciones",
+                  text: "Participa en emocionantes torneos y desafíos llenos de acción. ¡Compite, progresa y conviértete en una verdadera leyenda del juego!",
+                  link: "/competition",
+                  background: Competiciones, // Use the imported image
+                },
+                {
+                  icon: "fas fa-photo-video",
+                  title: "Contenido Exclusivo",
+                  text: "Sumérgete en el mundo de tu equipo favorito con contenido exclusivo: disfruta de entrevistas a tus ídolos, entrenamientos y momentos únicos que solo los verdaderos fanáticos pueden apreciar. ",
+                  link: "/content",
+                  background: Contenido, // Use the imported image
+                },
+                {
+                  icon: "fas fa-tasks",
+                  title: "Recompensas",
+                  text: "Disfruta completando tareas entretenidas mientras acumulas puntos con cada logro. Cuantos más desafíos superes, más puntos obtendrás para canjear por recompensas increíbles.",
+                  link: "/missions",
+                  background: Misiones, // Use the imported image
+                },
+              ].map((item, index) => (
+                <div
+                  key={index}
+                  className="relative flex flex-col items-center justify-center h-[30rem] w-[30rem] text-center sm:p-10 bg-[#f5f4f4] rounded-2xl backdrop-blur-md hover:shadow-2xl hover:scale-105 transition duration-300"
+                  style={{
+                    backgroundImage: item.background ? `url(${item.background})` : "none", // Apply background if available
+                    backgroundSize: "cover",
+                    backgroundPosition: "center"
+                  }}
+                >                
+                <div className="absolute inset-0 bg-black bg-opacity-50 rounded-2xl"></div> {/* Dark overlay */}
+                  <a href={item.link} className="w-full h-full absolute"></a>
+                  <i className={`${item.icon} text-[white] text-5xl sm:text-6xl mb-6 sm:mb-8 relative z-10`}></i>
+                  <p
+                    className="text-xl sm:text-2xl lg:text-[35px] font-semibold mb-4 sm:mb-6 relative z-10"
+                  >
+                    {item.title}
+                  </p>
+                </div>
+              ))}
+            </div>
+          </div>
+        </div>
+    )
+}
+
+export default FuncionaInicio;
